+++ spec/active_record_compose/inner_model_collection_spec.rb
@@ -32,4 +32,26 @@ RSpec.describe ActiveRecordCompose::InnerModelCollection do
       expect(collection).to be_empty
     end
   end
+
+  describe '#delete' do
+    subject(:collection) { ActiveRecordCompose::InnerModelCollection.new }
+
+    let(:account) { Account.new }
+    let(:profile) { Profile.new }
+
+    specify 'can be made empty by #clear' do
+      collection << account << profile
+      expect(collection.first).to eq account
+      collection.delete(account)
+      expect(collection.first).to eq profile
+    end
+
+    specify 'context must also be the same to get a hit' do
+      expect(collection).to be_blank
+      collection.push(account, context: :save)
+      expect(collection).to be_present
+      collection.delete(account, context: :destroy)
+      expect(collection).to be_present
+    end
+  end
 end